fix(media): add media files for Prompt, Chat, Character

+++ packages/frontend/src/store/FilesManager.ts
@@ -44,6 +44,13 @@ export class FilesManager {
   }
 
   async removeItem(id: string) {
+    if (this.temp[id]) {
+      delete this.temp[id];
+      runInAction(() => {
+        delete this.cache[id];
+      });
+      return;
+    }
     if (globalSettings.isBackendEnabled) {
       await backendManager.apiRequest({
         method: "DELETE",
